+++ src/Application/Framework/ProblemDetails/InternalServerErrorProblemDetails.php
@@ -12,24 +12,15 @@ class InternalServerErrorProblemDetails extends ProblemDetails
      * InternalServerErrorProblemDetails constructor.
      *
      * @param RequestInterface $request
-     * @param string|null      $reference
      */
-    public function __construct(RequestInterface $request, string $reference = null)
+    public function __construct(RequestInterface $request)
     {
-        $additional = null;
-        if (is_string($reference)) {
-            $additional = [
-                'reference' => $reference,
-            ];
-        }
-
         parent::__construct(
             '/problems/application/internal-server-error',
             'Internal Server Error',
             'An unknown error occurred.',
             500,
-            $request->getUri()->toRelative(),
-            $additional
+            $request->getUri()->toRelative()
         );
     }
 }