+++ src/Logger/Framework/Http/Middleware/Logger/LoggerMiddleware.php
@@ -7,8 +7,6 @@ use ExtendsSoftware\ExaPHP\Http\Middleware\Chain\MiddlewareChainInterface;
 use ExtendsSoftware\ExaPHP\Http\Middleware\MiddlewareInterface;
 use ExtendsSoftware\ExaPHP\Http\Request\RequestInterface;
 use ExtendsSoftware\ExaPHP\Http\Response\ResponseInterface;
-use ExtendsSoftware\ExaPHP\Logger\Exception\ReferencedException;
-use ExtendsSoftware\ExaPHP\Logger\Exception\ReferencedExceptionInterface;
 use ExtendsSoftware\ExaPHP\Logger\LoggerInterface;
 use ExtendsSoftware\ExaPHP\Logger\Priority\Error\ErrorPriority;
 use Throwable;
@@ -26,24 +24,19 @@ class LoggerMiddleware implements MiddlewareInterface
 
     /**
      * @inheritDoc
-     * @throws ReferencedExceptionInterface
+     * @throws Throwable
      */
     public function process(RequestInterface $request, MiddlewareChainInterface $chain): ResponseInterface
     {
         try {
             return $chain->proceed($request);
         } catch (Throwable $exception) {
-            $reference = uniqid();
-
             $this->logger->log(
                 $exception->getMessage(),
-                new ErrorPriority(),
-                [
-                    'reference' => $reference,
-                ]
+                new ErrorPriority()
             );
 
-            throw new ReferencedException($exception, $reference);
+            throw $exception;
         }
     }
 }